Fix final proptest by relaxing the constraint *again*.
Because the random data may be compressible before duplication we might have
a copy running from the end of one fragment and into the beginning of the
next, spanning the duplicate boundary.

This would then mean that the final copy would *not* be at least as long as
the string-we-duplicate.

Instead, just ensure that the final duplicate in the input is entirely covered
by copies in the output.